Update dashboards to be centralised
